coded astar movement and implemented it for rerr

+++ Assets/Scripts/Enemies/Enemy.cs
@@ -7,7 +7,7 @@ public abstract class Enemy : MonoBehaviour
     // fields
     protected float health;
     protected int defense;
-    protected int speed;
+    protected float speed;
     protected int damage;
     protected int knockback;
     protected float attackCooldown;
